refactor: move fixture to __fixtures__
